# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ChatMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TextContent"
                    },
                    {
                      "$ref": "#/components/schemas/ImageContent"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "title": "Content"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessageRequest",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "image_processor": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "auto",
            "title": "Image Processor"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageRequest"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "model_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "template": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template"
          },
          "tokenizer": {
            "default": "auto",
            "maxLength": 200,
            "minLength": 1,
            "title": "Tokenizer",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "messages"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "EvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Artifact Id",
            "type": "string"
          },
          "dataset_snapshot_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Dataset Snapshot Id",
            "type": "string"
          },
          "model_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "predict": {
            "default": true,
            "title": "Predict",
            "type": "boolean"
          },
          "template": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template"
          }
        },
        "required": [
          "model_id",
          "artifact_id",
          "dataset_snapshot_id"
        ],
        "title": "EvaluationRequest",
        "type": "object"
      },
      "ExportRequest": {
        "additionalProperties": false,
        "properties": {
          "adapter_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Adapter Id",
            "type": "string"
          },
          "model_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "node_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Node Id",
            "type": "string"
          },
          "output_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Output Name",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "adapter_id",
          "node_id",
          "output_name"
        ],
        "title": "ExportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageContent": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Asset Id",
            "type": "string"
          },
          "type": {
            "const": "image",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "asset_id"
        ],
        "title": "ImageContent",
        "type": "object"
      },
...
      "TextContent": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 100000,
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "text"
        ],
        "title": "TextContent",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OIDCBearer": {
        "bearerFormat": "OIDC JWT",
        "description": "Company OIDC access token. X-Project-ID is required for users with multiple projects.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Multimodal Fine-tuning MVP Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat/completions": {
      "post": {
        "operationId": "chat_api_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Chat Api V1 Chat Completions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "MVP result service"
        ]
      }
    },
...
    "/api/v1/evaluations": {
      "post": {
        "operationId": "evaluate_api_v1_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evaluate Api V1 Evaluations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate",
        "tags": [
          "MVP result service"
        ]
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Evaluation Api V1 Evaluations  Evaluation Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "MVP result service"
        ]
      }
    },
    "/api/v1/exports": {
      "post": {
        "operationId": "export_adapter_api_v1_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Adapter Api V1 Exports Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Adapter",
        "tags": [
          "MVP result service"
        ]
      }
    },
    "/api/v1/exports/{export_id}": {
      "get": {
        "operationId": "get_export_api_v1_exports__export_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Export Api V1 Exports  Export Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Export",
        "tags": [
          "MVP result service"
        ]
      }
    },
...
    "/api/v1/training-jobs/{job_id}/artifacts/{artifact_id}/download": {
      "get": {
        "operationId": "download_adapter_api_v1_training_jobs__job_id__artifacts__artifact_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Adapter",
        "tags": [
          "MVP result service"
        ]
      }
    },
...
    "/api/v1/training-jobs/{job_id}/manifest": {
      "get": {
        "operationId": "training_manifest_api_v1_training_jobs__job_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Training Manifest Api V1 Training Jobs  Job Id  Manifest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Training Manifest",
        "tags": [
          "MVP result service"
        ]
      }
    },
...
  },
  "security": [
    {
      "OIDCBearer": []
    }
  ],
...
}